+++ valores_players.csv
